+++ .gitignore
@@ -1,2 +1,4 @@
 UserInterfaceState.xcuserstate
 .DS_Store
+Pods/
+*.xcuserstate
