+++ tg/MainScreenListOfGames/ScreenAboutGame/ ScreenshotsFullScreen/ScreenshotViewer.swift
@@ -13,7 +13,7 @@ class ScreenshotViewer: UIViewController {
     var arrayOfScreenshots: [UIImage] = []
     var activeScreenshot = 0
     override func viewWillAppear(_ animated: Bool) {
-        SettingsViewController.applyUserSettings(currentClass: self, table: nil, collection: nil, searchController: nil, tableForHide: nil)
+        UserSettingsRegistration.apply(currentClass: self, table: nil, collection: nil, searchController: nil, tableForHide: nil)
     }
     override func viewDidLoad() {
         super.viewDidLoad()
