+++ tg/Settings/SettingsViewController.swift
@@ -19,7 +19,7 @@ class SettingsViewController: UIViewController {
         super.viewDidLoad()
         registerTableSettings()
         userPhoto.layer.cornerRadius = 15
-        SettingsViewController.applyUserSettings(currentClass: self, table: tableSettings, collection: nil, searchController: nil, tableForHide: nil)
+        UserSettingsRegistration.apply(currentClass: self, table: tableSettings, collection: nil, searchController: nil, tableForHide: nil)
     }
     func registerTableSettings() {
         tableSettings.delegate = self
@@ -27,36 +27,6 @@ class SettingsViewController: UIViewController {
         tableSettings.register(UINib(nibName: "TableSettingsCell", bundle: nil), forCellReuseIdentifier: "TableSettingsCell")
         tableSettings.separatorColor = .clear
         tableSettings.allowsSelection = false
-    }
-    static func applyUserSettings(currentClass: AnyObject, table: UITableView?, collection: UICollectionView?, searchController: UISearchController?, tableForHide: UITableView?) {
-        let nameOfStateArray = ["viewState", "themeState", "backgroundState"]
-        for name in nameOfStateArray {
-            let value = UserDefaults.standard.integer(forKey: name)
-            switch (name, value) {
-            case (nameOfStateArray[0], 0):
-                tableForHide?.isHidden = false
-            case (nameOfStateArray[0], 1):
-                tableForHide?.isHidden = true
-            case (nameOfStateArray[1], 0):
-                searchController?.searchBar.barTintColor = .systemGray6
-                currentClass.navigationController?.navigationBar.barTintColor = .systemGray6
-                currentClass.tabBarController??.tabBar.barTintColor = .systemGray6
-            case (nameOfStateArray[1], 1):
-                searchController?.searchBar.barTintColor = .darkGray
-                currentClass.navigationController?.navigationBar.barTintColor = .black
-                currentClass.tabBarController?.tabBar.barTintColor = .black
-            case (nameOfStateArray[2], 0):
-                currentClass.view?.backgroundColor = .white
-                table?.backgroundColor = .white
-                collection?.backgroundColor = .white
-            case (nameOfStateArray[2], 1):
-                currentClass.view?.backgroundColor = .lightGray
-                table?.backgroundColor = .lightGray
-                collection?.backgroundColor = .lightGray
-            default:
-                break
-            }
-        }
     }
 }
 
@@ -91,9 +61,6 @@ extension SettingsViewController: TableSettingsCellDelegate {
         default:
             break
         }
-        SettingsViewController.applyUserSettings(currentClass: self, table: tableSettings, collection: nil, searchController: nil, tableForHide: nil)
+        UserSettingsRegistration.apply(currentClass: self, table: tableSettings, collection: nil, searchController: nil, tableForHide: nil)
     }
 }
-
-
-
